pyproject.toml
[project]
name = "codereview-mcp"
version = "0.1.0"
description = "MCP server for AI-powered code review via OpenRouter"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.1",
    "openai>=1.0.0",
    "detect-secrets>=1.5.0",
]

[project.scripts]
codereview-mcp = "codereview_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]
